add Subcommand

+++ src/main/kotlin/events/CommandUtils.kt
@@ -38,19 +38,14 @@ class CommandFactory {
     fun onMessageEvent(event: MessageReceivedEvent) {
         if (event.author.isBot) return
         event.guild.getData()
-        var cont = true
         event.guild.punishments().forEach { punishment ->
             if (punishment != null && punishment.userId == event.author.id) {
-                try {
+                if (event.textChannel.canTalk()) {
                     event.message.delete().reason("This user is muted").queue()
-                    cont = false
-                } catch (ignored: Exception) {
+                    return
                 }
             }
-        }
-        messagesReceived.getAndIncrement()
-        if (cont) {
-            val member = event.member
+            messagesReceived.getAndIncrement()
             var args = event.message.rawContent.split(" ").toMutableList()
             val prefix = event.guild.getPrefix()
             when {
@@ -83,12 +78,13 @@ class CommandFactory {
                     } else {
                         executor.execute {
                             try {
-                                val data = member.data()
+                                val data = event.member.data()
                                 data.gold += 2
                                 data.update()
                                 cmd.executeInternal(args, event)
                                 r.table("commands").insert(r.json(getGson().toJson(
-                                        LoggedCommand(cmd.name, event.author.id, System.currentTimeMillis(), System.currentTimeMillis().readableDate())))).runNoReply(conn)
+                                        LoggedCommand(cmd.name, event.author.id, System.currentTimeMillis(), System.currentTimeMillis().readableDate())))
+                                ).runNoReply(conn)
                             } catch (e: Throwable) {
                                 e.log()
                                 event.channel.send("There was an exception while trying to run this command. Please join {0} and share the following stacktrace:".tr(event.guild, "<https://ardentbot.com/support>") + "\n${ExceptionUtils.getStackTrace(e)}")
@@ -102,39 +98,57 @@ class CommandFactory {
     }
 }
 
+abstract class Subcommand(val englishIdentifier: String, val syntax: String, val description: String? = null) {
+    abstract fun execute(arguments: MutableList<String>, event: MessageReceivedEvent)
+}
+
 abstract class Command(val category: Category, val name: String, val description: String, vararg val aliases: String) {
-    val help = mutableListOf<Pair<String, String>>()
+    val subcommands = mutableListOf<Subcommand>()
     fun executeInternal(args: MutableList<String>, event: MessageReceivedEvent) {
         if (event.channelType == ChannelType.PRIVATE)
             event.author.openPrivateChannel().queue { channel ->
                 channel.send("Please use commands inside a Discord server!".tr(Languages.ENGLISH.language))
             }
-        else execute(args, event)
+        else {
+            subcommands.forEach {
+                val identifier = it.englishIdentifier.tr(event.guild)
+                if (args.concat().startsWith(identifier)) {
+                    it.execute(args.concat().removePrefix(identifier).split(" ").toMutableList(), event)
+                    return
+                }
+            }
+            executeNoArgs(args, event)
+        }
     }
 
-    abstract fun execute(arguments: MutableList<String>, event: MessageReceivedEvent)
-
-    fun withHelp(syntax: String, description: String, event: MessageReceivedEvent): Command {
-        help.add(Pair(syntax, description.tr(event)))
+    fun withSubcommand(subcommand: Subcommand): Command {
+        subcommands.add(subcommand)
         return this
     }
 
-    fun displayHelp(channel: TextChannel, member: Member) {
+    abstract fun registerSubcommands()
+    abstract fun executeNoArgs(arguments: MutableList<String>, event: MessageReceivedEvent)
+
+    fun showHelp(event: MessageReceivedEvent) {
+        val member = event.member
+        val channel = event.textChannel
         val prefix = member.guild.getPrefix()
         val embed = member.embed("How can I use {0}?".tr(channel.guild).trReplace(channel.guild, "$prefix$name"), Color.BLACK)
                 .setThumbnail("https://upload.wikimedia.org/wikipedia/commons/f/f6/Lol_question_mark.png")
                 .setFooter("Aliases: {0}".tr(channel.guild).trReplace(channel.guild, aliases.toList().stringify()), member.user.avatarUrl)
                 .appendDescription("*${description.tr(channel.guild)}*\n")
-        help.forEach { embed.appendDescription("\n${Emoji.SMALL_BLUE_DIAMOND}**${it.first}**: *${it.second}*") }
-        if (help.size > 0) embed.appendDescription("\n\n**Example**: {0}".tr(channel.guild).trReplace(channel.guild, "$prefix$name ${help[0].first}"))
+        subcommands.forEach {
+            embed.appendDescription("\n" + Emoji.SMALL_BLUE_DIAMOND + "**" + it.syntax.tr(channel) + "**: "
+                    + (it.description?.tr(channel) ?: "No description is available for this command".tr(channel)))
+        }
+        if (subcommands.size > 0) embed.appendDescription("\n\n**Example**: {0}".tr(channel.guild).trReplace(channel.guild, "$prefix$name ${subcommands[0].syntax.tr(channel)}"))
         embed.appendDescription("\n\nType {0}help to view a full list of commands".tr(channel.guild).trReplace(channel.guild, member.guild.getPrefix()))
         channel.send(embed)
-        help.clear()
     }
 
     fun containsAlias(arg: String, guild: Guild): Boolean {
         val a = arg.split(" ")[0]
-        return a == name.tr(guild) || name.equals(a, true) || aliases.contains(a) || aliases.any {a == it }
+        return a == name.tr(guild) || name.equals(a, true) || aliases.contains(a) || aliases.any { a == it }
     }
 
     override fun toString(): String {
